refactor: replace getTotalCount with getPaginationInfo for improved pagination handling (#6306)
